+++ src/main/java/org/example/yardflow/repository/Registro_check_in_outRepositorio.java
@@ -10,7 +10,7 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface Registro_check_in_outRepositorio extends JpaRepository<Registro_check_in_out, Integer> {
+public interface Registro_check_in_outRepositorio extends JpaRepository<Registro_check_in_out, Long> {
 
 
     List<Registro_check_in_out> findAllByEntradapatio(LocalDate entradapatio);
@@ -19,13 +19,13 @@ public interface Registro_check_in_outRepositorio extends JpaRepository<Registro
     List<Registro_check_in_out> findAllBySaidapatio(LocalDate saidapatio);
 
     // Encontra o registro em aberto (entrada sem saída) mais recente da moto
-    Optional<Registro_check_in_out> findTopByMoto_IdmotoAndSaidapatioIsNullOrderByEntradapatioDesc(int idmoto);
+    Optional<Registro_check_in_out> findTopByMoto_IdmotoAndSaidapatioIsNullOrderByEntradapatioDesc(long idmoto);
 
     // Encontra o último registro (fechado ou não) pela moto
-    Optional<Registro_check_in_out> findTopByMoto_IdmotoOrderByEntradapatioDesc(int idmoto);
+    Optional<Registro_check_in_out> findTopByMoto_IdmotoOrderByEntradapatioDesc(long idmoto);
 
     // Lista todos os registros de uma moto
-    List<Registro_check_in_out> findByMoto_Idmoto(int idmoto);
+    List<Registro_check_in_out> findByMoto_Idmoto(long idmoto);
 
     // Lista todos com entrada (útil para ordenar por permanência em memória)
     List<Registro_check_in_out> findByEntradapatioIsNotNull();
